@@ -40,8 +40,16 @@ dependencies:
   shared_preferences: ^2.0.13
   device_info_plus: ^3.2.1
   flutter_local_notifications: ^9.2.0
+  permission_handler: ^9.0.1
   connectivity_plus: ^2.2.0
   rxdart: ^0.27.3
+  url_launcher: ^6.0.20
+  flutter_slidable: ^1.2.0
+  lazy_load_scrollview: ^1.3.0
+  page_transition: ^2.0.5
+  animated_splash_screen: ^1.2.0
+  flutter_mailer: ^2.0.1
+  qr_flutter: ^4.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -49,8 +57,6 @@ dependencies:
   provider: ^6.0.2
   flutter_sms: ^2.3.2
   font_awesome_flutter: ^9.2.0
-  qr_code_tools: ^0.0.7
-  razorpay_flutter: ^1.2.9
   image_picker: ^0.8.4+7
 
 dev_dependencies:
