@@ -1,19 +1,20 @@
 import 'package:animated_splash_screen/animated_splash_screen.dart';
 import '../Models/register_index_model.dart';
 import '../Modules/shared_preference_handler.dart';
-import 'package:chronicle/globalClass.dart';
-import 'package:chronicle/Widgets/googleSignInButton.dart';
-import 'package:connectivity/connectivity.dart';
+import '../global_class.dart';
+import '../Widgets/google_signin_button.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:page_transition/page_transition.dart';
 import '../Modules/database.dart';
-import '../customColors.dart';
-import 'MasterPages/chronicleMasterPage.dart';
-import 'errorDisplayPage.dart';
+import '../custom_colors.dart';
+import 'dashboard.dart';
+import 'error_display_page.dart';
 
 class RoutingPage extends StatefulWidget {
+  const RoutingPage({Key? key}) : super(key: key);
+
   @override
   _RoutingPageState createState() => _RoutingPageState();
 }
@@ -22,8 +23,7 @@ class _RoutingPageState extends State<RoutingPage> {
   GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey =
       GlobalKey<ScaffoldMessengerState>();
   Future<Widget> getWidget() async {
-    Widget widget;
-    List<RegisterIndexModel> lastRegisterModels;
+    Widget? widget;
     // initiateDatabase();
     Connectivity connectivity = Connectivity();
     await connectivity.checkConnectivity().then((value) async => {
@@ -32,93 +32,42 @@ class _RoutingPageState extends State<RoutingPage> {
               GlobalClass.user = FirebaseAuth.instance.currentUser,
               if (GlobalClass.user != null)
                 {
+                  print(1),
                   await addUserDetail().then((value) async => {
                         if (value != null)
                           {
-                            if (GlobalClass.userDetail.isOwner == 1 ||
-                                GlobalClass.userDetail.isAppRegistered == 1)
+                            if (GlobalClass.userDetail!.isOwner == 1 ||
+                                GlobalClass.userDetail!.isAppRegistered == 1)
                               await getLastRegister()
                                   .then((lastRegister) async => {
-                                        GlobalClass.lastRegister = lastRegister,
-                                        if (lastRegister == null ||
-                                            lastRegister == "")
-                                          {
-                                            widget = MaterialApp(
-                                                title: 'Chronicle',
-                                                debugShowCheckedModeBanner:
-                                                    false,
-                                                theme: lightThemeData,
-                                                darkTheme: darkThemeData,
-                                                themeMode: ThemeMode.system,
-                                                home: ChronicleMasterPage())
-                                          }
-                                        else
-                                          {
-                                            await getAllRegisterIndex()
-                                                .then((registers) => {
-                                                      GlobalClass.registerList =
-                                                          registers,
-                                                      lastRegisterModels =
-                                                          registers
-                                                              .where((element) =>
-                                                                  element
-                                                                      .id.key ==
-                                                                  lastRegister)
-                                                              .toList(),
-                                                      // sendNotifications(scaffoldMessengerKey,GlobalClass.userDetail.messageString),
-                                                      if (lastRegisterModels != null &&
-                                                          lastRegisterModels.length >
-                                                              0)
-                                                        widget = MaterialApp(
-                                                            title: 'Chronicle',
-                                                            debugShowCheckedModeBanner:
-                                                                false,
-                                                            theme:
-                                                                lightThemeData,
-                                                            darkTheme:
-                                                                darkThemeData,
-                                                            themeMode: ThemeMode
-                                                                .system,
-                                                            home: ChronicleMasterPage(
-                                                                register:
-                                                                    lastRegisterModels
-                                                                        .first))
-                                                      else
-                                                        widget = MaterialApp(
-                                                            title: 'Chronicle',
-                                                            debugShowCheckedModeBanner:
-                                                                false,
-                                                            theme:
-                                                                lightThemeData,
-                                                            darkTheme:
-                                                                darkThemeData,
-                                                            themeMode: ThemeMode
-                                                                .system,
-                                                            home:
-                                                                ChronicleMasterPage())
-                                                    })
-                                          }
+                                        widget = MaterialApp(
+                                            title: 'IKEP Buddy',
+                                            debugShowCheckedModeBanner: false,
+                                            theme: lightThemeData,
+                                            darkTheme: darkThemeData,
+                                            themeMode: ThemeMode.system,
+                                            home: DashboardPage())
                                       })
                             else
                               {
                                 widget = MaterialApp(
-                                    title: 'Chronicle',
+                                    title: 'IKEP Buddy',
                                     debugShowCheckedModeBanner: false,
                                     theme: lightThemeData,
                                     darkTheme: darkThemeData,
                                     themeMode: ThemeMode.system,
-                                    home: ChronicleMasterPage())
+                                    home: DashboardPage())
                               }
                           }
                         else
                           {
                             widget = MaterialApp(
-                                title: 'Chronicle',
+                                title: 'IKEP Buddy',
                                 debugShowCheckedModeBanner: false,
                                 theme: lightThemeData,
                                 darkTheme: darkThemeData,
                                 themeMode: ThemeMode.system,
-                                home: ErrorDisplayPage(
+                                home: const ErrorDisplayPage(
                                   appBarText: "Id Blocked",
                                   asset: "idBlocked.jpg",
                                   message:
@@ -130,7 +79,7 @@ class _RoutingPageState extends State<RoutingPage> {
               else
                 {
                   widget = MaterialApp(
-                      title: 'Chronicle',
+                      title: 'IKEP Buddy',
                       debugShowCheckedModeBanner: false,
                       theme: lightThemeData,
                       darkTheme: darkThemeData,
@@ -161,9 +110,9 @@ class _RoutingPageState extends State<RoutingPage> {
                                             // height: 400,
                                           ),
                                         ),
-                                        SizedBox(height: 20),
+                                        const SizedBox(height: 20),
                                         // Text(
-                                        //   'Chronicle',
+                                        //   'IKEP Buddy',
                                         //   style: TextStyle(
                                         //     color: CustomColors.firebaseYellow,
                                         //     fontSize: 40,
@@ -193,124 +142,19 @@ class _RoutingPageState extends State<RoutingPage> {
           else
             {
               widget = MaterialApp(
-                  title: 'Chronicle',
+                  title: 'IKEP Buddy',
                   debugShowCheckedModeBanner: false,
                   theme: lightThemeData,
                   darkTheme: darkThemeData,
                   themeMode: ThemeMode.system,
-                  home: ErrorDisplayPage(
+                  home: const ErrorDisplayPage(
                     appBarText: "No Internet Connection",
                     asset: "NoInternetError.webp",
                     message: 'Please connect to the Internet!!',
                   ))
             }
         });
-
-    // GlobalClass.user = FirebaseAuth.instance.currentUser;
-    // if(GlobalClass.user!=null){
-    //   await addUserDetail().then((value)async=>{
-    //     if(value!=null){
-    //       await getLastRegister().then((lastRegister) async => {
-    //         GlobalClass.lastRegister=lastRegister,
-    //         if(lastRegister==null||lastRegister=="")
-    //           {
-    //             widget=MaterialApp(
-    //                 title: 'Chronicle',
-    //                 debugShowCheckedModeBanner: false,
-    //                 theme: lightThemeData,
-    //                 darkTheme: darkThemeData,
-    //                 themeMode: ThemeMode.system,
-    //                 home: RegistersPage())
-    //           }
-    //         else {
-    //             await getAllRegisterIndex().then((registers) => {
-    //                 GlobalClass.registerList = registers,
-    //               lastRegisterModels=registers.where((element) => element.id.key==lastRegister).toList(),
-    //                 // sendNotifications(scaffoldMessengerKey,GlobalClass.userDetail.messageString),
-    //               if(lastRegisterModels!=null&&lastRegisterModels.length>0)widget=MaterialApp(
-    //                   title: 'Chronicle',
-    //                   debugShowCheckedModeBanner: false,
-    //                   theme: lightThemeData,
-    //                   darkTheme: darkThemeData,
-    //                   themeMode: ThemeMode.system,
-    //                   home: ClientPage(lastRegisterModels.first))
-    //               else widget=MaterialApp(
-    //                   title: 'Chronicle',
-    //                   debugShowCheckedModeBanner: false,
-    //                   theme: lightThemeData,
-    //                   darkTheme: darkThemeData,
-    //                   themeMode: ThemeMode.system,
-    //                   home: RegistersPage())
-    //             })
-    //         }
-    //       })
-    //       }
-    //     else{
-    //       widget= MaterialApp(
-    //       title: 'Chronicle',
-    //       debugShowCheckedModeBanner: false,
-    //       theme: lightThemeData,
-    //       darkTheme: darkThemeData,
-    //       themeMode: ThemeMode.system,
-    //       home: ErrorDisplayPage(appBarText: "Id Blocked",asset: "idBlocked.jpg",message: 'Please contact System Administrator',))
-    //     }
-    //   });
-    // }
-    // else {
-    //   widget= MaterialApp(
-    //       title: 'Chronicle',
-    //       debugShowCheckedModeBanner: false,
-    //       theme: lightThemeData,
-    //       darkTheme: darkThemeData,
-    //       themeMode: ThemeMode.system,
-    //       home: ScaffoldMessenger(key: scaffoldMessengerKey,child: Scaffold(body: SafeArea(
-    //         child: Padding(
-    //           padding: const EdgeInsets.only(
-    //             left: 16.0,
-    //             right: 16.0,
-    //             bottom: 20.0,
-    //           ),
-    //           child: Column(
-    //             mainAxisSize: MainAxisSize.max,
-    //             children: [
-    //               Row(),
-    //               Expanded(
-    //                 child: Column(
-    //                   mainAxisSize: MainAxisSize.min,
-    //                   mainAxisAlignment: MainAxisAlignment.center,
-    //                   children: [
-    //                     Flexible(
-    //                       child: Image.asset(
-    //                         'assets/firebase_logo.png',
-    //                         // height: 400,
-    //                       ),
-    //                     ),
-    //                     SizedBox(height: 20),
-    //                     // Text(
-    //                     //   'Chronicle',
-    //                     //   style: TextStyle(
-    //                     //     color: CustomColors.firebaseYellow,
-    //                     //     fontSize: 40,
-    //                     //   ),
-    //                     // ),
-    //                     // Text(
-    //                     //   'Authentication',
-    //                     //   style: TextStyle(
-    //                     //     color: CustomColors.firebaseBlue,
-    //                     //     fontSize: 40,
-    //                     //   ),
-    //                     // ),
-    //                   ],
-    //                 ),
-    //               ),
-    //               GoogleSignInButton(scaffoldMessengerKey: scaffoldMessengerKey,)
-    //             ],
-    //           ),
-    //         ),
-    //       ),),));
-    // }
-
-    return widget;
+    return widget!;
   }
 
   @override
@@ -321,7 +165,7 @@ class _RoutingPageState extends State<RoutingPage> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-        title: 'Chronicle',
+        title: 'IKEP Buddy',
         debugShowCheckedModeBanner: false,
         theme: lightThemeData,
         darkTheme: darkThemeData,
