@@ -1,9 +1,13 @@
 import 'package:flutter/cupertino.dart';
 
 class ModalOptionModel {
-  final String? particulars;
-  final Function? onTap;
-  final IconData? icon;
+  final String particulars;
+  final void Function() onTap;
+  final IconData icon;
   final Color? iconColor;
-  ModalOptionModel({this.particulars, this.onTap, this.icon, this.iconColor});
+  ModalOptionModel(
+      {required this.particulars,
+      required this.onTap,
+      required this.icon,
+      this.iconColor});
 }
